% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios822FinalProjPackages.R
\name{preprocess_flowdata}
\alias{preprocess_flowdata}
\title{Preprocess a flowframe}
\usage{
preprocess_flowdata(fdat_raw, col_name)
}
\arguments{
\item{fdat_raw}{(flowFrame) raw flow data}

\item{col_name}{the markers name for preprocessing}
}
\value{
fdat_trans (flowFrame) flow data after compensation & arcsinh transformation
}
\description{
compensate the input raw flow data and transform
the compensated data using arcsinh transformation
}
