% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bios822FinalProjPackages.R
\name{read_preprocess_flowdata}
\alias{read_preprocess_flowdata}
\title{Preprocess a flowframe by inputting file name}
\usage{
read_preprocess_flowdata(file_name, col_name)
}
\arguments{
\item{file_name}{(str) fcs file name}
}
\value{
fdat_trans (flowFrame) preprocessed fcs data
}
\description{
a wrapper function of preprocess_flowdata
}
